small updates to domain bridging and system interface file structure
